removed IKZ label from submodules
