+++ IKZ_plugin/src/czochralski/parser.py
@@ -30,23 +30,23 @@ from nomad.datamodel.data import (
 )
 
 from nomad_material_processing.utils import create_archive
-from ds_IKZ import DirectionalSolidificationExperiment
+from czochralski import MeltCzochralskiExperiment, Sensors
 
-class RawFileDigitalProtocol(EntryData):
+class CSVFile(EntryData):
     measurement = Quantity(
-        type=DirectionalSolidificationExperiment,
+        type=Sensors,
         a_eln=ELNAnnotation(
             component='ReferenceEditQuantity',
         )
     )
 
 
-class DSParserIKZ(MatchingParser):
+class CzParser(MatchingParser):
 
     def __init__(self):
         super().__init__(
-            name='Directional Solidification IKZ',
-            code_name= 'Directional Solidification IKZ',
+            name='Czochralski IKZ',
+            code_name= 'Czochralski IKZ',
             code_homepage='https://github.com/FAIRmat-NFDI/AreaA-data_modeling_and_schemas',
             supported_compressions=['gz', 'bz2', 'xz']
         )
@@ -54,9 +54,9 @@ class DSParserIKZ(MatchingParser):
     def parse(self, mainfile: str, archive: EntryArchive, logger) -> None:
         data_file = mainfile.split('/')[-1]
         data_file_with_path = mainfile.split("raw/")[-1]
-        entry = DirectionalSolidificationExperiment()
-        entry.digital_protocol_file = data_file_with_path
-        file_name = f'{data_file[:-4]}.archive.json'
+        entry = Sensors()
+        entry.data_file = data_file_with_path
+        file_name = f'{data_file[:-12]}.archive.json'
         #entry.normalize(archive, logger)
-        archive.data = RawFileDigitalProtocol(measurement=create_archive(entry,archive,file_name))
-        archive.metadata.entry_name = data_file + ' growth file'
+        archive.data = CSVFile(measurement=create_archive(entry,archive,file_name))
+        archive.metadata.entry_name = data_file + ' measurement file'
