+++ IKZ_plugin/src/directional_solidification/nomad_plugin.yaml
