+++ IKZ_plugin/src/directional_solidification/parser.py
@@ -18,7 +18,6 @@
 
 from nomad.datamodel import EntryArchive
 from nomad.metainfo import (
-    Section,
     MSection,
     Quantity,
 )
@@ -31,31 +30,23 @@ from nomad.datamodel.data import (
 )
 
 from nomad_material_processing.utils import create_archive
-from movpe_IKZ import SubstrateInventory
-from basesections_IKZ import (
-    IKZMOVPECategory
-)
+from directional_solidification import DirectionalSolidificationExperiment
 
-class RawFileSubstrateInventory(EntryData):
-    m_def = Section(
-        a_eln=None,
-        categories=[IKZMOVPECategory],
-        label = 'Raw File Substrate Inventory'
-    )
+class RawFileDigitalProtocol(EntryData):
     measurement = Quantity(
-        type=SubstrateInventory,
+        type=DirectionalSolidificationExperiment,
         a_eln=ELNAnnotation(
             component='ReferenceEditQuantity',
         )
     )
 
 
-class MovpeSubstrateParser(MatchingParser):
+class DSParserIKZ(MatchingParser):
 
     def __init__(self):
         super().__init__(
-            name='MOVPE Substrate IKZ',
-            code_name= 'MOVPE Substrate IKZ',
+            name='Directional Solidification IKZ',
+            code_name= 'Directional Solidification IKZ',
             code_homepage='https://github.com/FAIRmat-NFDI/AreaA-data_modeling_and_schemas',
             supported_compressions=['gz', 'bz2', 'xz']
         )
@@ -63,9 +54,9 @@ class MovpeSubstrateParser(MatchingParser):
     def parse(self, mainfile: str, archive: EntryArchive, logger) -> None:
         data_file = mainfile.split('/')[-1]
         data_file_with_path = mainfile.split("raw/")[-1]
-        entry = SubstrateInventory()
-        entry.substrate_data_file = data_file_with_path
-        file_name = f'{data_file[:-5]}.archive.json'
+        entry = DirectionalSolidificationExperiment()
+        entry.digital_protocol_file = data_file_with_path
+        file_name = f'{data_file[:-4]}.archive.json'
         #entry.normalize(archive, logger)
-        archive.data = RawFileSubstrateInventory(measurement=create_archive(entry,archive,file_name))
-        archive.metadata.entry_name = data_file + ' substrates file'
+        archive.data = RawFileDigitalProtocol(measurement=create_archive(entry,archive,file_name))
+        archive.metadata.entry_name = data_file + ' growth file'
