--- IKZ_plugin/src/mbe_IKZ/__init__.py
@@ -1,2 +1,0 @@
-from .schema import *
-from basesections_IKZ import *
