+++ IKZ_plugin/src/movpe/movpe1_growth_parser/deposition_control/parser.py
@@ -39,7 +39,7 @@ from nomad.processing.data import Upload
 from nomad.app.v1.models.models import User
 from nomad.datamodel.data import EntryData, ArchiveSection
 from nomad.search import search, MetadataPagination
-from movpe_IKZ import (
+from movpe import (
     ExperimentMovpe1IKZ,
     GrowthMovpe1IKZConstantParametersReference,
     GrowthMovpe1IKZDepositionControl,
@@ -64,7 +64,7 @@ from nomad.datamodel.datamodel import EntryArchive, EntryMetadata
 # from nomad.parsing.tabular import create_archive
 from nomad.utils import hash
 
-from basesections_IKZ import IKZMOVPE1Category
+from basesections import IKZMOVPE1Category
 
 
 def create_archive(
